smart-librarian/requirements.txt
openai
chromadb
tiktoken
streamlit
python-dotenv
